requirements.txt
fastapi==0.110.1
alembic==1.13.1
asyncpg==0.29.0
SQLAlchemy==2.0.29
pydantic==2.7.0
python-dotenv==1.0.1
uvicorn==0.29.0